Update: using float32 for shifts in utils/random/functions

+++ src/utils/random/functions.py
@@ -13,8 +13,10 @@ def random_shifted_functions(src_fn: tff.TensorflowFunction,
                              hshift_bounds: typing.Tuple[float, float]) -> \
       typing.List[tff.TensorflowFunction]:
   rng = np.random.default_rng()
-  vshifts = rng.uniform(vshift_bounds[0], vshift_bounds[1], n)
-  hshifts = rng.uniform(hshift_bounds[0], hshift_bounds[1], n)
+  vshifts = rng.uniform(vshift_bounds[0],
+                        vshift_bounds[1], n).astype(np.float32)
+  hshifts = rng.uniform(hshift_bounds[0],
+                        hshift_bounds[1], n).astype(np.float32)
 
   def transformed_fn(v: float, h: float):
     nonlocal src_fn
